+++ routes/api.php
@@ -38,8 +38,9 @@ Route::group(['middleware' => 'auth:api'], function () {
     Route::post('apikey/activatekey/{institution}', [ ApiKeyController::class, 'activateApiKey']); #activate apikey for external offers
 });
 
-#test route für pagination
-Route::get('pagination', [OfferController::class, 'paginate']);
+//test routes for pagination
+Route::get('offer/paginated/{offerCount}', [OfferController::class, 'paginated_offers']);
+Route::get('list/offer/short/paginated/{offerCount}', [ OfferController::class, 'paginated_reduced_offers']);
 
 // Additional routes that don't require authentication
 Route::get('offer', [ OfferController::class, 'index']);